fixed2 -> float2

+++ Assets/GPGPUSandbox/CubeWave/Shaders/Kernels.shader
@@ -11,26 +11,26 @@
 
     sampler2D _PositionBuffer;
 
-    fixed2 random2(fixed2 st)
+    float2 random2(float2 st)
     {
-        st = fixed2(dot(st, fixed2(127.1, 311.7)),
-                    dot(st, fixed2(269.5, 183.3)));
+        st = float2(dot(st, float2(127.1, 311.7)),
+                    dot(st, float2(269.5, 183.3)));
         return -1.0 + 2.0 * frac(sin(st) * 43758.5453123);
     }
 
-    float perlin_noise(fixed2 st) 
+    float perlin_noise(float2 st) 
     {
-        fixed2 p = floor(st);
-        fixed2 f = frac(st);
-        fixed2 u = f * f * (3.0 - 2.0 * f);
+        float2 p = floor(st);
+        float2 f = frac(st);
+        float2 u = f * f * (3.0 - 2.0 * f);
 
-        float v00 = random2(p + fixed2(0, 0));
-        float v10 = random2(p + fixed2(1, 0));
-        float v01 = random2(p + fixed2(0, 1));
-        float v11 = random2(p + fixed2(1, 1));
+        float v00 = random2(p + float2(0, 0));
+        float v10 = random2(p + float2(1, 0));
+        float v01 = random2(p + float2(0, 1));
+        float v11 = random2(p + float2(1, 1));
 
-        return lerp(lerp(dot(v00, f - fixed2(0, 0)), dot(v10, f - fixed2(1, 0) ), u.x),
-                    lerp(dot(v01, f - fixed2(0, 1)), dot(v11, f - fixed2(1, 1) ), u.x), 
+        return lerp(lerp(dot(v00, f - float2(0, 0)), dot(v10, f - float2(1, 0) ), u.x),
+                    lerp(dot(v01, f - float2(0, 1)), dot(v11, f - float2(1, 1) ), u.x), 
                     u.y) + 0.5f;
     }
     
@@ -44,7 +44,7 @@
     float4 frag_update_position(v2f_img i) : SV_Target
     {
         float4 p = tex2D(_PositionBuffer, i.uv);
-        p.y = perlin_noise(fixed2(i.uv.x * 10, i.uv.y * 10 + _Time.x * 10)) * 10;
+        p.y = perlin_noise(float2(i.uv.x * 10, i.uv.y * 10 + _Time.x * 10)) * 10;
         return p;
     }
     ENDCG
